move AgaveClient out into its own file
